CAL-94 moved security classes to new module name video-security
