+++ catalog/video/video-app/src/main/resources/features.xml
@@ -48,7 +48,7 @@
         <bundle>mvn:org.codice.alliance/stanag4609/${project.version}</bundle>
         <bundle>mvn:org.codice.alliance/mpegts/${project.version}</bundle>
         <bundle>mvn:org.codice.alliance/klv/${project.version}</bundle>
-        <bundle>mvn:org.codice.alliance.platform/security/${project.version}</bundle>
+        <bundle>mvn:org.codice.alliance.video/video-security/${project.version}</bundle>
     </feature>
 
 </features>
