+++ catalog/video/video-mpegts-stream/pom.xml
@@ -123,8 +123,8 @@
             <version>${commons-collections4.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.codice.alliance.platform</groupId>
-            <artifactId>security</artifactId>
+            <groupId>org.codice.alliance.video</groupId>
+            <artifactId>video-security</artifactId>
             <version>${project.version}</version>
         </dependency>
