--- platform/security/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-/**
- * Copyright (c) Codice Foundation
- *
- * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
- * version 3 of the License, or any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
- * <http://www.gnu.org/licenses/lgpl.html>.
- *
- **/
- -->
-<blueprint xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.1.0"
-           xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0">
-
-    <bean id="videographerClaimsHandler" class="org.codice.alliance.security.claims.videographer.VideographerClaimsHandler">
-
-        <cm:managed-properties persistent-id="org.codice.alliance.security.claims.videographer.VideographerClaimsHandler"
-                               update-strategy="container-managed"/>
-
-        <property name="attributes">
-            <list value-type="java.lang.String">
-                <value>http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier=videographer</value>
-                <value>http://schemas.xmlsoap.org/ws/2005/05/identity/claims/role=videographer</value>
-            </list>
-        </property>
-    </bean>
-
-    <service ref="videographerClaimsHandler" interface="org.apache.cxf.sts.claims.ClaimsHandler"/>
-
-    <bean id="videographerValidator" class="org.codice.alliance.security.validator.videographer.VideographerValidator">
-
-       <cm:managed-properties persistent-id="org.codice.alliance.security.validator.videographer.VideographerValidator"
-                               update-strategy="container-managed"/>
-
-        <property name="supportedRealms">
-            <list>
-                <value>karaf</value>
-                <value>ldap</value>
-            </list>
-        </property>
-    </bean>
-
-    <service ref="videographerValidator" interface="org.apache.cxf.sts.token.validator.TokenValidator"/>
-
-</blueprint>
