+++ pom.xml
@@ -471,7 +471,6 @@
     </pluginRepositories>
     <modules>
         <module>support</module>
-        <module>platform</module>
         <module>catalog</module>
         <module>distribution</module>
         <module>libs</module>
